backend/requirements.txt
# Core Django & REST Framework
asgiref
Django
djangorestframework
djangorestframework-simplejwt  # For JWT authentication

# Database
mysqlclient

# Utilities
sqlparse
Pillow  # For image handling
django-cors-headers  # For handling CORS

requests
reportlab
